Tagx00.MachineLearning/requirements.txt
jieba
tensorflow
numpy
flask
baidu-aip